expose metrics prompts by renaming metric modules
fixes: #208
